Final commit.
Configured to connect with ngrok webservice handling http requests and a
heart rate monitor + medicine injection setup. Will not deploy correctly
unless these services are disabled or rerouted. Our final work, as
presented!

+++ HellowApp6/Service1.cs
@@ -22,7 +22,7 @@ namespace HellowApp6
     {
         private Patient _patient;
         private int i;
-        private List<int> BPM = new List<int>();
+        private List<double> BPM = new List<double>();
         public override IBinder OnBind(Intent intent)
         {
             return new Service1Binder(this);
@@ -62,7 +62,7 @@ namespace HellowApp6
     public class Vital
     {
         public string meaning { get; set; }
-        public int value { get; set; }
+        public double value { get; set; }
         public object received { get; set; }
     }
 }
